fix: Fix reverse list type
