+++ py-polars/tests/unit/series/test_series.py
@@ -1862,27 +1862,6 @@ def test_cumulative_eval() -> None:
     assert_series_equal(s.cumulative_eval(expr3), expected3)
 
 
-def test_reverse() -> None:
-    s = pl.Series("values", [1, 2, 3, 4, 5])
-    assert s.reverse().to_list() == [5, 4, 3, 2, 1]
-
-    s = pl.Series("values", ["a", "b", None, "y", "x"])
-    assert s.reverse().to_list() == ["x", "y", None, "b", "a"]
-
-
-def test_reverse_binary() -> None:
-    # single chunk
-    s = pl.Series("values", ["a", "b", "c", "d"]).cast(pl.Binary)
-    assert s.reverse().to_list() == [b"d", b"c", b"b", b"a"]
-
-    # multiple chunks
-    chunk1 = pl.Series("values", ["a", "b"])
-    chunk2 = pl.Series("values", ["c", "d"])
-    s = chunk1.extend(chunk2).cast(pl.Binary)
-    assert s.n_chunks() == 2
-    assert s.reverse().to_list() == [b"d", b"c", b"b", b"a"]
-
-
 def test_clip() -> None:
     s = pl.Series("foo", [-50, 5, None, 50])
     assert s.clip(1, 10).to_list() == [1, 5, None, 10]
